+++ examples/starter_header_footer_and_menu.py
@@ -46,4 +46,4 @@ layout = Layout([
     ])
 ])
 
-quasargui.run(layout, debug=False)
+quasargui.run(layout, debug=True)