+++ quasargui/__init__.py
@@ -6,5 +6,5 @@ this_directory = dirname(__file__)
 QUASAR_GUI_ASSETS_PATH = join(this_directory, 'assets')
 QUASAR_GUI_INDEX_PATH = join(QUASAR_GUI_ASSETS_PATH, 'index.html')
 
-from .main import run, set_main_component
-from ._base import Model
+from quasargui.main import run, set_main_component
+from quasargui.base import Model