+++ quasargui/form.py
@@ -1,8 +1,7 @@
-from typing import Union
-
-from quasargui._base import ComponentWithModel, Component, Model
-from quasargui._tools import build_props
-from quasargui.typing import ClassesType, StylesType, PropsType, EventsType
+from quasargui.base import *
+from quasargui.base import Component
+from quasargui.tools import build_props
+from quasargui.typing import *
 
 
 class Input(ComponentWithModel):
@@ -25,6 +24,7 @@ class Input(ComponentWithModel):
 
 
 class Button(Component):
+    component = 'q-btn'
     defaults = {
         'props': {
             'unelevated': True,
@@ -34,7 +34,7 @@ class Button(Component):
     def __init__(self,
                  label: str = None,
                  icon: str = None,
-                 color: Union[Model, str] = None,
+                 color: PropValueType[str] = None,
                  classes: ClassesType = None,
                  styles: StylesType = None,
                  props: PropsType = None,
@@ -45,8 +45,20 @@ class Button(Component):
             'color': color})
         super().__init__(classes=classes, styles=styles, props=props, events=events)
 
-    @property
-    def vue(self) -> dict:
-        return self._merge_vue({
-            'component': 'q-btn'
-        })
+
+class Toggle(ComponentWithModel):
+    """
+    ref. https://quasar.dev/vue-components/toggle#qtoggle-api
+    """
+    component = 'q-toggle'
+
+    def __init__(self,
+               label: str = None,
+               model: Model = None,
+               classes: ClassesType = None,
+               styles: StylesType = None,
+               props: PropsType = None,
+               events: EventsType = None):
+        props = build_props({}, props, {
+            'label': label})
+        super().__init__(model=model, classes=classes, styles=styles, props=props, events=events)